refactor: Moved `Response` to `Responses/Response`
