+++ tests/Feature/MockRequestTest.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-use Saloon\Http\Response;
 use Saloon\Helpers\MockConfig;
 use Saloon\Http\PendingRequest;
 use League\Flysystem\Filesystem;
 use Saloon\Http\Faking\MockClient;
+use Saloon\Http\Responses\Response;
 use Saloon\Http\Faking\MockResponse;
 use Saloon\Exceptions\Request\RequestException;
 use Saloon\Tests\Fixtures\Requests\UserRequest;
