+++ tests/Fixtures/Data/User.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Saloon\Tests\Fixtures\Data;
 
-use Saloon\Http\Response;
+use Saloon\Http\Responses\Response;
 
 class User
 {
