+++ tests/Fixtures/Responses/CustomResponse.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Saloon\Tests\Fixtures\Responses;
 
-use Saloon\Http\Response;
+use Saloon\Http\Responses\Response;
 
 class CustomResponse extends Response
 {
