+++ tests/Fixtures/Responses/UserResponse.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Saloon\Tests\Fixtures\Responses;
 
-use Saloon\Http\Response;
+use Saloon\Http\Responses\Response;
 
 class UserResponse extends Response
 {
