+++ tests/Unit/ConfigTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-use Saloon\Http\Response;
 use Saloon\Helpers\Config;
 use Saloon\Http\PendingRequest;
 use Saloon\Http\Faking\MockClient;
+use Saloon\Http\Responses\Response;
 use Saloon\Http\Faking\MockResponse;
 use Saloon\Http\Senders\GuzzleSender;
 use Saloon\Tests\Fixtures\Senders\ArraySender;
