+++ tests/Unit/MockResponseTest.php
@@ -2,8 +2,8 @@
 
 declare(strict_types=1);
 
-use Saloon\Http\Response;
 use Saloon\Http\Faking\MockClient;
+use Saloon\Http\Responses\Response;
 use Saloon\Http\Faking\MockResponse;
 use Saloon\Tests\Fixtures\Responses\UserData;
 use Saloon\Tests\Fixtures\Requests\UserRequest;
